refactor(handlers): align SpaceStation imports with WarShip topology
Extract ISpaceStation and the SpaceStationType enum into a new sibling
leaf module SpaceStationUnitHandler.types.ts so the .calculations.ts
and .helpers.ts leaves no longer back-import the orchestrator. This
mirrors WarShip, where the type lives outside the orchestrator and
the leaves are pure (never importing from .ts).

Closes the remaining 2 of 4 circular dependencies in
src/services/units/handlers/:
  - SpaceStationUnitHandler.calculations.ts <-> SpaceStationUnitHandler.ts
  - SpaceStationUnitHandler.helpers.ts <-> SpaceStationUnitHandler.ts

Public API preserved: ISpaceStation and SpaceStationType are
re-exported from SpaceStationUnitHandler.ts (and through the
index.ts barrel) so existing consumers and tests compile unchanged.
BV parity verified via scripts/validate-bv.ts: per-unit results
byte-identical pre/post-refactor.
